pyproject.toml
[project]
name = "sekimiya-ai-langchain"
version = "0.1.0"
description = "Let's create your very own Sekimyia!"
authors = [
    { name = "lune-sta", email = "lune@wuvu.net" }
]
dependencies = [
    "langchain>=0.0.27",
    "openai>=0.27.8",
    "discord.py>=2.3.1",
    "tiktoken>=0.4.0",
    "google-api-python-client>=2.95.0",
    "pyowm>=3.3.0",
    "setuptools>=68.0.0",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
